Revert to subclassing
