+++ testsuite/MDAnalysisTests/topology/test_mol2.py
@@ -27,7 +27,7 @@ from numpy.testing import (
 
 import MDAnalysis as mda
 
-from MDAnalysisTests.topology.test_base import ParserBase
+from MDAnalysisTests.topology.base import ParserBase
 from MDAnalysisTests.datafiles import (
     mol2_molecule,
     mol2_molecules,
