+++ testsuite/MDAnalysisTests/topology/test_pdb.py
@@ -29,7 +29,7 @@ from numpy.testing import (
 )
 import MDAnalysis as mda
 
-from MDAnalysisTests.topology.test_base import ParserBase
+from MDAnalysisTests.topology.base import ParserBase
 from MDAnalysisTests.datafiles import (
     PDB,
     PDB_small,
